Radically changed network interface handlers API: first of all, we now speak of "network interfaces" instead of "interfaces", for the single reason that "NETIF" is then a logical typedef whereas "IF" is simply not allowed because it's apparantly an element of the C language ;)
As with the LED driver API, the "LED" structure disappears completely. All
knowledge related to this belongs in the main program. The only thing that
network interface handlers actually still need to know about is the LED state,
since they decide in what color the LED should light up.

And likewise, there was a need for a seperation between the interface handler
itself and the state information that is now kept inside a NETIF structure
for handler-internal use (it remains opaque to the main program). This structure
also sort of replaces previous ifh_pdata and ifh_cdata fields.

powerctl() and colorctl() have been replaced by a single col() function that
determines power and color at the same time by setting a LEDSTATE variable
accordingly.


git-svn-id: file:///home/pief/svn/rleds/trunk@23 822f2e9a-261e-0410-a646-d699a81b6f75